LayoutTests/inspector/worker/runtime-basic-expected.txt

Test for RuntimeAgent in a Worker.


== Running test suite: Worker.Runtime.basic
-- Running test case: Main.Runtime.evaluate
Target - Page - passphrase - page-passphrase

-- Running test case: Worker.Runtime.evaluate
Target - TestWorker - passphrase - worker-passphrase

-- Running test case: Main.Runtime.RemoteObjectAndPropertyDescriptor.
Target - Page - location and href - Location: inspector/worker/runtime-basic.html

-- Running test case: Worker.Runtime.RemoteObjectAndPropertyDescriptor.
Target - TestWorker - location and href - WorkerLocation: inspector/worker/resources/worker-1.js
